feat: deploy from seed, linting, ERC721

+++ contracts/interfaces/IWETH10.sol
@@ -15,7 +15,7 @@ import "./IERC3156FlashLender.sol";
 /// balance of ETH deposited minus the ETH withdrawn with that specific wallet.
 interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
     /// @dev Returns current amount of flash-minted WETH10 token.
-    function flashMinted() external view returns(uint256);
+    function flashMinted() external view returns (uint256);
 
     /// @dev `msg.value` of ETH sent to this contract grants caller account a matching increase in WETH10 token balance.
     /// Emits {Transfer} event to reflect WETH10 token mint of `msg.value` from `address(0)` to caller account.
@@ -26,7 +26,7 @@ interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
     function depositTo(address to) external payable;
 
     /// @dev Burn `value` WETH10 token from caller account and withdraw matching ETH to the same.
-    /// Emits {Transfer} event to reflect WETH10 token burn of `value` to `address(0)` from caller account. 
+    /// Emits {Transfer} event to reflect WETH10 token burn of `value` to `address(0)` from caller account.
     /// Requirements:
     ///   - caller account must have at least `value` balance of WETH10 token.
     function withdraw(uint256 value) external;
@@ -44,7 +44,11 @@ interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
     /// Requirements:
     ///   - `from` account must have at least `value` balance of WETH10 token.
     ///   - `from` account must have approved caller to spend at least `value` of WETH10 token, unless `from` and caller are the same account.
-    function withdrawFrom(address from, address payable to, uint256 value) external;
+    function withdrawFrom(
+        address from,
+        address payable to,
+        uint256 value
+    ) external;
 
     /// @dev `msg.value` of ETH sent to this contract grants `to` account a matching increase in WETH10 token balance,
     /// after which a call is executed to an ERC677-compliant contract with the `data` parameter.
@@ -58,9 +62,13 @@ interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
     /// Emits {Approval} event.
     /// Returns boolean value indicating whether operation succeeded.
     /// For more information on {approveAndCall} format, see https://github.com/ethereum/EIPs/issues/677.
-    function approveAndCall(address spender, uint256 value, bytes calldata data) external returns (bool);
+    function approveAndCall(
+        address spender,
+        uint256 value,
+        bytes calldata data
+    ) external returns (bool);
 
-    /// @dev Moves `value` WETH10 token from caller's account to account (`to`), 
+    /// @dev Moves `value` WETH10 token from caller's account to account (`to`),
     /// after which a call is executed to an ERC677-compliant contract with the `data` parameter.
     /// A transfer to `address(0)` triggers an ETH withdraw matching the sent WETH10 token in favor of caller.
     /// Emits {Transfer} event.
@@ -68,5 +76,9 @@ interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
     /// Requirements:
     ///   - caller account must have at least `value` WETH10 token.
     /// For more information on {transferAndCall} format, see https://github.com/ethereum/EIPs/issues/677.
-    function transferAndCall(address to, uint value, bytes calldata data) external returns (bool);
+    function transferAndCall(
+        address to,
+        uint256 value,
+        bytes calldata data
+    ) external returns (bool);
 }
